[PROJECT] [UPDATE] Refactor CloudfrontService; migrate to edge.service and remove BucketService
