+++ src/routes/project/project.controller.spec.ts
@@ -5,7 +5,7 @@ import { PrismaService } from "@ourPrisma/prisma.service";
 import { ConfigService } from "@nestjs/config";
 import { S3Service } from "@s3/s3.service";
 import { BucketService } from "@s3/bucket.service";
-import { CloudfrontService } from "@s3/cloudfront.service";
+import { CloudfrontService } from "src/routes/s3/edge.service";
 
 describe("ProjectController", () => {
   let controller: ProjectController;
